src/main/java/com/quest/RestartServlet.java
package com.quest;

import jakarta.servlet.ServletException;
import jakarta.servlet.annotation.WebServlet;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.IOException;

@WebServlet(name = "restart", value = "/restart")
public class RestartServlet extends HttpServlet{
    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        req.setCharacterEncoding("UTF-8");
        resp.setCharacterEncoding("UTF-8");
        String answer = req.getParameter("answer");
        if ("yes5".equals(answer)) {
            req.getRequestDispatcher("/index.jsp").forward(req, resp);
        }
    }
}
